refactor: split trade insights ai leniency helpers
